+++ CHANGELOG.md
@@ -16,3 +16,21 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/)
 ### Fixed
 
 ### Removed
+
+## [v0.0.1](https://github.com/labforward/config/releases/tag/v0.0.1) - 2021-05-06
+
+### Added
+
+- [[#CUB-3561]](https://labforward.atlassian.net/browse/CUB-3561) New package @labforward/config
+- [[#CUB-3561]](https://labforward.atlassian.net/browse/CUB-3561) New package @labforward/eslint-config-base
+- [[#CUB-3561]](https://labforward.atlassian.net/browse/CUB-3561) New package @labforward/eslint-config-cypress
+- [[#CUB-3561]](https://labforward.atlassian.net/browse/CUB-3561) New package @labforward/eslint-config-node
+- [[#CUB-3561]](https://labforward.atlassian.net/browse/CUB-3561) New package @labforward/eslint-config-react
+
+### Changed
+
+### Deprecated
+
+### Fixed
+
+### Removed